requirements.txt
django
djangorestframework
requests
--upgrade pip